Move the health probe and repair budget to platform/health.rs
The esphome config probe, its temp dir handling, and the repair attempt
counter that bounds what a failed probe may trigger. The counter helpers
move with it since health owns the budget; ensure_user_python's defer
counting imports them back. PROBE_TIMEOUT becomes pub(super) for
interpreter_is_usable, which stays with the tree lifecycle code.
